Move UI-free log_store/search_history under tui/views
Both old and new view trees share one copy.
