+++ src/tui/views/log_store.rs
