+++ src/tui/views/mod.rs
@@ -1,0 +1,5 @@
+pub mod log_store;
+pub mod search_history;
+
+pub use log_store::{LogEntry, LogStore};
+pub use search_history::SearchHistory;
