+++ src/tui/views/search_history.rs
